requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography~=41.0.3
Deprecated==1.2.14
idna==3.4
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
pip~=23.2.1
wheel~=0.41.2
docutils~=0.20.1
setuptools~=68.1.2